begin overhaul of settings
