+++ app/src/main/res/layout/fragment_reports.xml
@@ -10,13 +10,11 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        android:layout_marginEnd="8dp"
-        android:layout_marginStart="8dp"
         android:layout_marginTop="16dp"
         android:fontFamily="@font/avenir_medium"
         android:text="@string/repots_title"
         android:textColor="@color/dukeBlue"
-        android:textSize="55sp"
+        android:textSize="@dimen/title"
         android:textStyle="bold"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
