+++ app/src/main/res/values/dimens.xml
@@ -2,4 +2,10 @@
     <!-- Default screen margins, per the Android Design guidelines. -->
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
+    <dimen name="title">40sp</dimen>
+    <dimen name="margin_top">40dp</dimen>
+    <dimen name="title_margin_bottom">20dp</dimen>
+    <dimen name="button_width">250dp</dimen>
+    <dimen name="text_size">25sp</dimen>
+    <dimen name="margin_start">25dp</dimen>
 </resources>
